Check for a permission to modify course

+++ courses/views.py
@@ -1,7 +1,8 @@
 from django.shortcuts import get_object_or_404
 from rest_framework.views import APIView
 from rest_framework.response import Response
-from rest_framework import status
+from rest_framework import status, permissions
+from courses.permissions import IsCreatorOrReadOnly
 
 from courses.models import Course
 from courses.serializers import CourseSerializer
@@ -9,6 +10,11 @@ from courses.serializers import CourseSerializer
 
 class CourseList(APIView):
     """List all courses"""
+    permission_classes = [
+        permissions.IsAuthenticatedOrReadOnly,
+        IsCreatorOrReadOnly
+    ]
+
     def get(self, request):
         courses = Course.objects.all()
         serializer = CourseSerializer(courses, many=True)
@@ -26,6 +32,11 @@ class CourseList(APIView):
 
 class CourseDetail(APIView):
     """Create, get, update or delete a course"""
+    permission_classes = [
+        permissions.IsAuthenticatedOrReadOnly,
+        IsCreatorOrReadOnly
+    ]
+
     def get(self, request, pk):
         course = get_object_or_404(Course, pk=pk)
         serializer = CourseSerializer(course)
@@ -33,6 +44,7 @@ class CourseDetail(APIView):
 
     def put(self, request, pk):
         course = get_object_or_404(Course, pk=pk)
+        self.check_object_permissions(request, course)
         serializer = CourseSerializer(course, data=request.data)
 
         if serializer.is_valid():
@@ -42,5 +54,6 @@ class CourseDetail(APIView):
 
     def delete(self, request, pk):
         course = get_object_or_404(Course, pk)
+        self.check_object_permissions(request, course)
         course.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
